windowarrangementcuj: Get multiresizer reliably
Fixing a problem with how clamshell WindowArrangementCUJ tests get the
multiresizer. The multiresizer is summoned by hovering the mouse near
where two windows meet. As it may be summoned unintentionally, it is
dismissed by continuing to move the mouse around that area (without
moving it to the multiresizer). So the tests need to detect when the
multiresizer appears, and immediately stop moving the mouse, because any
further mouse movement may cause the multiresizer to disappear. If the
multiresizer disappears, the impact is sometimes that the test fails
with an error about not finding the multiresizer, but more often, the
test just proceeds with a "resizing" drag that doesn't really resize.

BUG=b:253294086
TEST=tast run <DUT> ui.WindowArrangementCUJ.clamshell_mode
TEST=tast run <DUT> ui.WindowArrangementCUJ.lacros

Change-Id: If2f8de2b628fc2fbff78bc6cc0378de8d71474af