Lacros variants for multivm memory tests
The Lifecycle memory tests create tabs, so they need to support Lacros
explicitly. Other tests don't interact with the browser, but it's
important to measure any memory differences between using Lacros and
not.

This CL adds a selection of Lacros variants to all the multivm tests
that measure memory.

memoryuser has been updated to take an optional browser.Type argument,
which is then used to create the new *browser.Browser field in TestEnv.

BUG=b:246896386
TEST=tast run dut multivm.Login.* multivm.Lifecycle*
multivm.MemoryChromeOSPerf.*

Change-Id: I73549f78ee98bc1979303e6c3ea97f48b4085b1f