tast-tests: Fix testAppFromOutsideOfPlayStore in arc.ResizeLock
testAppFromOutsideOfPlayStore is supposed to test the cases when an app
is installed outside of PlayStore, resize lock should not be enabled for
it. However, the current test is not testing those cases - the test
installs the app through PlayStore; and then launch it in maximized
mode(that is the reason why the test is passing currently, since resize
lock is not enabled for maximized windows).

Therefore, this change fixes the test to actually test what is intended
to be tested, by installing an app outside of PlayStore and launch it in
phone size.

Also note that we are observing flakiness in arc.ResizeLock related to
"failed to start activity", which seems to be due to the apps not being
installed successfully. One possible reason can be that at the beginning
of the test, we are installing five apps all at once. This commit also
moves the installation of the aforementioned phone size app into the
function of testAppFromOutsideOfPlayStore, as an attempt to address the
flakiness.

BUG=b:243452113
TEST=tast run arc.ResizeLock

Change-Id: I3452491845927f11f757e16fafc658db45310cc5