feedback: Verify if the system_logs.zip is in the report
This CL adds the test to verify if the system_logs.zip is in the report.
Both the histograms.zip and system_logs.zip should be included if the
checkbox is checked and should not be included if the checkbox is
unchecked.

BUG=b:185624798
TEST=tast run <DUT>
feedback.ChooseToShareSystemAndAppInfoAndMetricsOrNot.*

Change-Id: I011c8fddbff74452030c229c8d27ec1ccb6c9961
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3967304
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Danny Wang <wangdanny@google.com>
Commit-Queue: Danny Wang <wangdanny@google.com>
Reviewed-by: Wenyu Zhang <zhangwenyu@google.com>