wmp: Fix wmp.WindowCycle*.lacros tests
There is a current wait for the app to be shown. According to shelf.go,
it waits for appID shown on shelf. Since we are opening the same app on
multiple desks, it would be shown on the shelf after opening on the
first desk. Use a different waiter to check that a window pops up on the
active (newly created desk), and also let the open animation finish.

BUG=b:253039033
TEST=tast run <IP> wmp.WindowCycle*.lacros

Change-Id: If733efb6f97fe93d7a347673cb26d572edc04a6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3952978
Tested-by: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Hongyu Long <hongyulong@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Avery Musbach <amusbach@chromium.org>