cellular: Increase the time for ReAttach to 5 seconds
The test is failing on qualcomm devices because the ReAttach takes just
over 2 seconds, and the command to set the initial EPS bearer settings
fails because the modem is busy finishing the ReAttach procedure.
Increase the delay to give the modem time to stabilize.

BUG=b:249592531
TEST=Tested with Verizon and AT&T on limozeen

Change-Id: Id51d1999a3712ce30f05150f6bff263d3dc6cb85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3954462
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>