tast-tests: Add a betty variant for arc.OptinManaged
Adding a betty variant will help us add special treatment for betty
board because it has stability issues that do not appear on normal
boards.

In this change we're also dumping logcat to be able to investigate
provisioning related problems.

Reducing the scope of the test to waiting for mangaed provisioning to
completed instead of launching Play Store. That is already covered in
enterprise.ARCProvisioning test.

BUG=b:248290063
TEST=tast run DUT arc.OptinManaged.vm

Change-Id: Icf8590ca5a12bdd448d3832540a22d6bd014d311