tast-tests: Generate drive URL by model name
Currently, generating the random drive URLs is deterministic for each
sku by seeding our random number generator with the sku id. However, the
current method of finding that sku is faulty, because it assumes that
there is a file that exists that has the sku information, and that file
is not present in most of our devices. This change seeds the random
generator that selects the drive link with the model name instead, which
should definitely be accessible within the tast test.

BUG=b:253518019
TEST=tast run <DUT> ui.GoogleSlidesCUJ

Change-Id: I00d93403a730831a64a4d3b4771e16fe27c0fd82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3954762
Tested-by: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Dean Liao <deanliao@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>