arc: Not test toggling between Phone and Tablet in arc.ResizeLock for some models
For the models with a small display, when toggling from Phone to Tablet,
ARC falls back to maximize the app since the requested bounds from
Chrome is too large for the display. This causes the test of toggling
between Phone and Tablet failing.

This commit adds a hard-coded list for those models and skips the test
for them, as a quick workaround to address the test flakiness.

BUG=b:243452113
TEST=tast run arc.ResizeLock*

Change-Id: Iebf489a07cbabcb97977a6bc0fe90036448fc310
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3955985
Reviewed-by: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Reviewed-by: Georg Neis <neis@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Tested-by: Qijing Yao <yaoqq@google.com>
Commit-Queue: Qijing Yao <yaoqq@google.com>