perfutil+ui: Fix error handling of RunMultiple
perfutil.RunMultiple turned an error into a nil slice, thus losing
valuable error information. Moreover, callers of RunMultiple didn't
check for nil.

The CL fixes this and goes one step further: the only thing all call
sites do with the result of RunMultiple is call the Save method on it.
So we can make this part of RunMultiple -- renamed now to
RunMultipleAndSave. Moreover, since all callsites already have to pass
the testing state (s) for uiperf.Run, I think it's justified to also
pass it to RunMultipleAndSave and thus simplify things a bit.

BUG=chromium:1373942
TEST=fast_build.sh -T
TEST=tast run $DUT ui.OverviewScrollPerf
TEST=tast run $DUT ui.OverviewScrollPerf  # test closure modified to fail

Change-Id: I49d83b444d0ef9bc6330660a764d6937bdada0dd