tast-tests: Lacros variants to shelf.ActivityIndicators
This test needs to be migrated to Lacros in order to check the
activity indicator for any types of browser.

For Lacros PWA app test, Lacros would need to be set keep-alive to
install the PWA by policy by a new fixture. By the way, there seems to
be a race failing to wait for the app to be installed and accessible
randomly. To address this flake InstallPwaAppByPolicy will have retry
logic to refresh policy then try installing once more.

BUG=b:219815913, b:236325022
TEST=tast run $DUT shelf.ActivityIndicators*lacros

Change-Id: I55fbe6fa88449d3dc03cf634b272e2ca994e0897
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3954778
Tested-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Hyung Tae Kim <hyungtaekim@chromium.org>