tast-tests: Relaunch Play Store if it closes during verification
Tast board shows that one of the common failure reasons for
enterprise.ARCProvisioning.unstable is during verification of Play Store
empty step. Screenshot of the failed runs show that Play Store is closed
in all those runs. Logcat shows that Play Store is forcefully closed due
to update of GmsCore.

In this change, we'll relaunch Play Store if it closes.

BUG=b:242902484
TEST=tast run DUT enterprise.ARCProvisioning.vm_unstable

Change-Id: Ic58a16a3b0d930bfe2845714202cba32c8e45034