ui: Wait for overview mode to exit in DesksCUJ
This change ensures that the overview switcher in DesksCUJ waits until
overview mode has exited before checking the active desk index.

Prior to this change, after clicking on the desks mini views, we wait
until no event occurs at the root node. I was able to reproduce and
record the test failing for this reason once, and the recording showed
the device freezing up right before ending overview mode. I suspect
that when the device froze up, "WaitUntilNoEvent" completed
successfully, and then we checked for the active desk too early.
Regardless of if this change actually fixes the "unexpected active desk"
issue, I believe that we should still wait for the overview mode
animation to complete.

BUG=b:249157486
TEST=tast run ui.DesksCUJ*

Change-Id: Ia3a9a16f74fbc93a0f4a7eb33cca47a45c994938