[TAST] Refactor BTTestService to support Floss.
This change refactors the existing BTTestService to provide support for
Floss. The major changes required were:

- Move required BlueZ logic out of the service.
- Remove no longer required logic.
- Expand the generic Bluetooth interface.

This is possible due to the fact that the planned Bluetooth UI tests do
not use any lower level Bluetooth functionality, and instead will opt to
rely solely on information from the UI and from interacting with the UI.
This change does not restrict us from adding more in the future.

A convenience method that returns the names and addresses of Bluetooth
devices found is provided.

BUG=b:242057403,b:242071154
TEST=Manually on a local device and on lab devices.

Change-Id: I63c2cfb73d1d5c4d8e51dd357c5c00d8ee7c9338