tast-tests: Convert the add-remove factors tests over to using fixtures
This switches the tests over to using fixtures to set up the USS. It
also adds a separate variant of the fixture for ensuring the opposite,
that USS is disabled.

I also reworked some of the way that the fixture itself works, adding a
real return type for the SetUp that tests can then use to see if they're
running with a USS or no-USS config. This saves tests from having to
define their own separate parameter to track this.

BUG=b:246993770
TEST=tast run <dut> <these tests>

Change-Id: I0d1cc8bddd0f13fe4e347e6f6837353c74533086
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3949732
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Kamil Szarek <kamilszarek@google.com>
Reviewed-by: Chenghao Piao <cpiao@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>