Use account pool and fake dms to serve the policy; add retries
In order to stabilize this test, a fake policy server was used so there
is one less point of failure. Retries were also added for operations
that are not specific to the feature being tested. Verification of chrome log upload was removed since that work is done with DMS and outside the scope of what we need to test for this feature.

BUG=b:235603221
TEST=tast run dut enterprise.ARCInstallLogging.vm

Change-Id: Iac822332060d7ef03cd3a733a3ae168dac5f7be7