tast-tests: Add remove to the AddRemoveFactors test.
This (fairly obvious) portion of the test was not added when I
originally wrote it because the RemoveAuthFactor support was not quite
working yet, but it was always intended and so here it is.

Unfortunately, in the process of testing this I also discovered that
remove doesn't always seem to work correctly with smart cards and USS,
so I had to disable that path in that case again pending the discovered
bug being fixed.

BUG=b:254277646
TEST=tast run <dut> cryptohome.AddRemoveFactors.*

Change-Id: I977f887702fc1135bcb8789cb5f363baaa2b9339