Tast: Don't wait on Attract Loop launch in demomode.ChromeApps
Asserting that the Attract Loop app launches is extremely flaky
due to the length of time it takes the AL to download on the
first session after Demo Mode setup. Let's disable it for now to make
the test more stable while we investigate other sources of flakiness and
figure out what to do about the Attract Loop.

BUG=b:254311499
TEST=tast -verbose run <device-ip> demomode.ChromeApps

Change-Id: Ie0096960211ce6a75310f74a24e8f07c0260f1a5