wmp: Reduce flakiness in OverviewCloseAll
In this test we change the body of the testing.Poll function a bit to
give it a 2 second timeout, which reduces flakiness that could be caused
by windows taking time to be forced closed after the initial one second
delay that we give windows to close asynchronously. We also change the
s.Fatal return in the Poll to a PollBreak for better error handling
within the Poll.

BUG=b:243451406
TEST=tast run <IP> wmp.OverviewCloseAll

Change-Id: I6092e4506e152ee89431438be7030297e1f07edb