Make autotestPrivate usage in TestAPIConn optional
TestAPIConn, the constructor function for TestConn, calls an
autotestPrivate function for initialization. In Lacros we disallow
autotestPrivate, leading to some error messages in the Lacros log:

[extension_function_dispatcher.cc(674)] Unknown Extension API [...]
[CONSOLE(0)] "Unchecked runtime.lastError: Access to extension API [...]

Put this call behind a parameter flag and only do it where needed
(generally speaking: for Ash).

To achieve this, also make browser.Browser carry the browser type.

BUG=None
TEST=./fast_build.sh -T
TEST=manually with custom test on DUT, while inspecting lacros.log

Change-Id: I95ca53efd0cd7ebd2010d0b1dee258bcc14c9cac