fs_corruption: Don't error when restarting crostini fails
At the end of the test we try to put crostini back the way it was so
the fixture doesn't have to reset it, but this can fail. This
shouldn't be considered a test failure, as the feature being tested
still works and the test fixture will be able to repair things either
way. Further, the startup failures we see appear to be spurious
i.e. the crostini VM actually is running afterwards and the fixture
will be able to notice this.

BUG=b:253150565
TEST=none

Change-Id: I8c3a716740c6d63bc809c471638eaa045a2bf134