tast-tests: Restore image after rollback using fixture
Restoring the image using non-enterprise rollback fails
sometimes. The autoupdate fixture also includes the logic to
restore the image if it does not match the one that was in the
device beforehand, so we can rely on the fixture instead of
having repeated code.

Remove the method that restores the image from the rollback
helper.

Update the logs in the fixture to have a clear distinction
between the different cleanup steps and possible failures.
New logs may help in the future in case the fixture also fails
when restoring the image.

Add PreTest logic to make sure the test is run in the original
version.

BUG=b:241391509
TEST=tast run -var=updateutil.tlwAddress="localhost:7151" DUT
autoupdate.EnterpriseRollbackPreviousVersion.rollback_1_version

Change-Id: I41ba00bb14e1edc93e5b7a1eab9d7fadf80053bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3966688
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>