printer: Add hplip sublanguages
hplip has two filters (hpps and hpcups).  Within hpcups, there are 12
values of the hpPrinterLanguage sublanguage represented in our PPD
index.  This adds sub-tests of printer.Add that cover all of these.  The
two existing tests that cover ljcolor and hbpl1 are also renamed to
match the same convention.  This should allow us to catch regressions
that only affect a subset of HP languages.

pcl3 and quickconnect are limited to x86 because they currently produce
unexplained differences when run on arm.

BUG=b:236843006, b:236841459
TEST=tast run dut-grunt printer.Add.hp* printer.ProxyAdd.hp*

Change-Id: I64cc210cd73c8953557dadc51ede401e5d25d245