ServicesOnBoot.sound_card_init: Enable only by boot time calibration
There are three types of devices:
1.  sound_card_init does not exist on the board.
2.  sound_card_init exists for other models of the same board,
    but does nothing for the device as the device does not have a
    smart amp.
3.  sound_card_init exists and does boot time calibration for the
    smart amp.

Case 1 fails the test because `initctl status sound_card_init` fails.
Ideally we want the test to cover cases 2 and 3.
But given that we do not have a way to tell if the sound_card_init
service exists as a hardware or software dependency,
limiting the test to run for case 3 only is also fine,
as it's the only case where we care about sound_card_init running
successfully.

BUG=b:254556002
BUG=b:254566972
TEST=tast run platform.ServicesOnBoot.sound_card_init

Change-Id: I5747a646c8cfc2e0e7224c13c5a0cf6b4af1ff1f