lacros: ShelfLaunchOmaha finds compatible build combos from VersionHistory
There was a bug that ShelfLaunchOmaha couldn't parse the OS "beta"
and "stable" channels, which may have caused it to skip testing "beta"
and "stable" Lacros channels respectively.

This change will find the live versions of Lacros using VersionHistory
API by the time it runs.
It includes only the versions that are compatible with Ash.

Also it adds the "unstable" variant to cover arm64 devices.

BUG=chromium:1375175, chromium:1367120
TEST=tast run -checktestdeps=false ${DUT} lacros.ShelfLaunchOmaha  # ran on the OS stable, beta and dev

Change-Id: I75bd32379721b0ecda84d14acff6974777466750