oobe: Extend SmokeEndToEnd with the HPS screen for the eligible devices
Add new HPS test case to the oobe.SmokeEndToEnd test, refactor assistant
screen logic to handle new/existing assistant users differemtly as we
show a different amount of screens for them. Because we use the same
name for the skip button on both assistant and smart privacy screens
we need to explicitly click through the assistant to be able to verify
whether we show a correct screen for devices with the HPS enabled.
Add new fixtures to display OOBE starting from the welcome screen and
to show hps screen.

BUG=b:254653584
TEST=tast run {DUT_IP} oobe.SmokeEndToEnd
TEST=tast run {DUT_IP} oobe.SmokeEndToEnd.hps
TEST=tast unit tests passed.

Change-Id: I50e4cb299d29f651d5e7daae5fd46a44464030d5