perf: Add snapshot grace period in perf.Timeline
perf.Timeline currently fails if one snapshot takes longer than the
specified timeline interval. This change adds a 1-interval grace period
for the snapshot to complete. For example, if we collect data every 5
seconds, and one snapshot happens to take 6 seconds, we will skip the
next datapoint and log the error, but not fail the timeline. If the
snapshot takes longer than 2 total intervals (in this example, longer
than 10 seconds), the timeline collection will fail. This code will
limit the likelihood of getting bad data by failing in the case that a
single snapshot takes unnecessarily and unexpectedly long, while not
punishing tests that just barely miss the interval.

BUG=b:252829815
TEST=To force reproduction of the timeline interval errors, I added in a
sleep to the snapshot code that was equal to the 5s. I then ran
ui.QuickCheckCUJ, and verified that the timeline did not fail, and that
results-chart.json correctly showed the timing in the TPS.t metric. In
addition, I separately added a 10s sleep to a timeline with a 5s
interval, and verified that the test would properly fail.

Change-Id: I8f91f324a4368dbaef11988866ab89ca3f0e8042
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3969642
Commit-Queue: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Tested-by: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>