tast-tests: Add tast for VPN source routing
This patch adds a tast to verify that traffic from different users is
routed appropriately when a VPN is connected, either through the VPN or
out the underlying physical network.

This patch also contains a fix for allowing ping to run as users without
home directories.

BUG=b:254374443
TEST=tast run network.VPNSourceRouting
TEST=tast run network.VPNConnect

Change-Id: I4d9707b8bd5cf455e447f6099da669a1de835f5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3969898
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>