tast-tests: crostini: Reenable the DiskIoPerf test
This was disabled a long time ago due to flakiness, but there's no
record of what the error was. It ran fine against a random lab DUT and
we'd really like this coverage, especially for manatee, so reenable it.
If it's not still flaky, great. If it is, then we need the data to root
cause.
Also fix an unrelated typo because why not.

BUG=None
TEST=CQ

Change-Id: I9bae0c3af3fa3469d2f4ba536fd59ad3bd69ffe8