hps: wait for hpsd features at the right time
Waiting for hpsd to bind its D-Bus name *before* we enable any HPS
functionality serves little purpose, because at that point hpsd has just
powered the sensor back off.

Instead, *after* enabling HPS functionality, wait for hpsd to be fully
started and the corresponding sensor feature to be enabled. This
accounts for the ~6 seconds needed to power on the sensor, as well as
any time needed for flashing firmware if necessary.

If hpsd fails at any point in that process, this should give us a more
informative error faster.

This also lets us remove a delay which was trying to account for the ~6
second power on time in the sensor.

This also gives us a good point in the test where we know that the
sensor is powered on and running. A subsequent commit will use that to
check the running firmware version. Previously, the sensor wasn't
powered on until ~6 seconds after changing the settings, at which time
the tests would already be sleeping.

BUG=b:250660161
TEST=tast run -var=tablet=$TABLET $DUT 'hps.Camerabox*'

Change-Id: I0e98102535ebc7fa9e0393c10c2176409bef7ec7