hps: tighten up slack in expected screen dim durations
The previous commit gives us a better way to account for the ~6 seconds
hpsd takes to power on the HPS peripheral.

Conservatively reduce the slack from 10 seconds to 5 seconds (although
it seems empirically we could reduce this much closer to 0).

Doing so requires us to clarify when we start counting the time. The
hps.CameraboxLoLOn test is parametrized, with one parameter expecting
quick dim and the other expecting no quick dim. In the quick dim case we
need to allow the ~6 seconds power on time but in the other case we need
to exclude it.

BUG=b:250660161
TEST=tast run -var=tablet=$TABLET $DUT 'hps.Camerabox*'

Change-Id: Ie887d4a6e2bb93f7979a3ed37abd3f47c4116197