hps: tweak some wording to make hps.CameraboxLoLOff easier to follow
The hps.CameraboxLoLOff test case is a negative test, with the opposite
expectations to hps.CameraboxLoLOn, as you would expect from its name.
But the test description and commentary suggested it wasn't, and I got
confused when trying to understand what exactly the test was doing.

Tweak the wording in an attempt to clarify what's happening in this
test.

BUG=b:250660161
TEST=none

Change-Id: Ie1d16c7485e524a19875eaf74cc0e69e030e1cd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3970500
Reviewed-by: Alvin Jia <alvinjia@google.com>
Reviewed-by: Michael Pishchagin <mblsha@google.com>
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>