arc: Fix arc.PIPRoundedCornersUnderlay on kevin and scarlet
Apparently it started failing because the histogram is updated before
taking snapshot via metrics.Run.

This can be fixed by passing arcpipvideotest.EstablishARCPIPVideo to
metrics.Run so that the snapshot is taken before executing it.

Also replace metrics.Run with metrics.RunAndWaitAll with a 3-second
timeout to remove redundant testing.Sleep.

BUG=b:254744188
TEST=tast run <kevin> arc.PIPRoundedCornersUnderlay

Change-Id: I8c7a625ef1e9c3178d9932fd7ccba3f3c74a298c