tast-tests: Add fixup test case to arc.GhostWindow
This CL adds a test case for the fixup ghost window. The test case first
prepares for a long P-to-R fixup for the Play Store app by creating a
lot of directories in the Play Store's package-specific data directory
in SDCard partition
(/data/media/0/Android/data/com.android.vending), signs out and signs in
again, launches Play Store, and verifies that a ghost window is popped
up with a message starting from "Updating Android System".

This CL also adds a check that the non-fixup ghost windows display a
message "Starting <App Name>".

BUG=b:186767881
TEST=tast run <DUT> arc.GhostWindow.fixup_r
TEST=tast run <DUT> arc.GhostWindow.general_r
TEST=tast run <DUT> arc.GhostWindow.fullrestore_r

Change-Id: I37948bbb913ad746ae08da5cd40397b63e72fb8e