windowarrangementcuj: Initialize the window bounds
Sometimes the clamshell WindowArrangementCUJ tests fail with "failed to
resize browser window from the upper left to the middle and back"
because the expected window bounds don't account for the magnetism where
if you resize within 10 DIPs of the work area boundary, the resize snaps
to the work area boundary. To avoid this problem, the tests will
initialize the browser window bounds predictably and never so close to
the work area boundary.

BUG=b:253294086
TEST=tast run <DUT> ui.WindowArrangementCUJ.clamshell_mode
TEST=tast run <DUT> ui.WindowArrangementCUJ.lacros

Change-Id: I8f9208b203cb495746f5ea46c40fe5234a01b348