ui: Make toggling File menu in MeetCUJ more robust
After toggling the File menu, the test now waits until the ui tree
actually changes before proceeding to close it. Additionally, now we try
clicking on the File menu with a retry, to handle window focusing issues
on some devices.

In updating the File menu toggle, I also change moving the mouse to the
File menu to fail the test on error, as opposed to logging it. I do not
expect moving the mouse to fail unless the node does not exist, and thus
it would be important to fail the test and fix the failure rather than
let it silently continue.

BUG=b:254916317
TEST=tast run <DUT> ui.MeetCUJ.4p_present_notes_split

Change-Id: If7d9083bacb0246e335bd26bf5206c3ac246c56a