Add DisplayService in WWCB testing.
In the wwcb test, most of the fixtures need to be connected
to the chromebook through the dock. In order to verify the
situation of the external fixture, it is necessary to judge
by grabbing the system information of the chrombook, and the
display service can provide information about the screen and
verify the window on the ceratin display.

BUG=b:255496810
TEST=tast run <DUT> wwcb.Dock1PersistentGRPC

Change-Id: If02984053b626f5f0c05bebba9ccfdf859aec2f6