faft: Log disp bus info in ECLaptopMode
When the display turns on/off, the runtime usage count
changes with it. Log this information to help evaluate
whether it'd be a good alternative in checking for the
display's power state. Also, if setting power state with
servo fails, retry with sending a press on the pwr button.

BUG=b:200305492
TEST=manual on eldrid, storo360, and vilboz

Change-Id: Ic74d9e2e6a3e8079212370dd897efbf2ba9f8b4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3977677
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Arthur Chuang <arthur.chuang@cienet.com>
Commit-Queue: Arthur Chuang <arthur.chuang@cienet.com>