graphics: Prepare HTMLVideoRoundedCornersUnderlay
The test is mostly green and I want to promote it to critical, but there
are still a handful of models where it is failing and may not be fixed
quickly. Skip the tests on the failing models, and create "failing"
variants that run only on those models.

BUG=b:255612376
TEST=../tast/fast_build.sh -T

Change-Id: I063ed9d1fdb4e69931a6207f0a1696e698f7a595
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3980013
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Georg Neis <neis@chromium.org>