ui: Prepare ChromePIPRoundedCornersUnderlay
The test is mostly green and I want to promote it to critical, but there
are still a handful of models where it is failing and may not be fixed
quickly. Skip the tests on the failing models, and create "failing"
variants that run only on those models.

BUG=b:255612376
TEST=../tast/fast_build.sh -T

Change-Id: I8e180ed1ade2f1317df207f8c67e1605a5308f24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3980014
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Avery Musbach <amusbach@chromium.org>