ui: Create ExampleCUJ
ExampleCUJ serves as an illustrative example on how to use the CUJ
recorder. It performs keyboard and mouse events to trigger their
respective input latency metrics and interacts with the Ash UI to
collect meaningful ADF metrics. Additionally, it navigates back and
forth between two websites to demo collecting FCP and LCP metrics.

BUG=b:254145320
TEST=tast run <DUT> ui.ExampleCUJ ui.ExampleCUJ.lacros

Change-Id: Ia4f08f52e4c41d4511622e2281a811efb717a1cd