feedback: Combine tast test cases
This cl combines previous tast test cases into one test. Description
should always be included in the report. Screenshot should be included
in the report when the screenshot checkbox is checked. Diagnostic data
should be included in the report when the Diagnostic data checkbox is
checked. Email should be included in the report by default. Attachment
should be included in the report when the attach file checkbox is
checked.

BUG=b:185624798
TEST=tast run <DUT> feedback.VerifyEssentialData.*

Change-Id: Ice363e867236bae335a2afbdbef010e7455cfcd0