hps: check that HPS is running the expected firmware
Cross-check the firmware version reported by the HPS peripheral against
the version number bundled with the firmware we have on disk, to ensure
they match.

This prepares for subsequent commits, where we introduce new tests which
run different firmware. Checking the version will help us to ensure we
are testing the firmware we are intending to test.

BUG=b:252908633
TEST=tast run -var=tablet=$TABLET $DUT hps.DBus 'hps.Camerabox*'
TEST=tast run $VM hps.DBus.fake

Change-Id: Icf0a79fee5fe050bb2cf1730e1fd879838d6b045