tast-tests: Update network.ConfigBaseline
OobeNetworkConfigBaseline checks the netconfig API works during OOBE.

Rename and extend test to also check the API works after login. There
are complex tests (like rollback) that rely on this API to work. The
idea is adding this tests to the CQ to be able to detect issues in the
network API timely.

BUG=b:250876596
TEST=tast run DUT network.ConfigBaseline.oobe
tast run DUT network.ConfigBaseline.login

Change-Id: I035129ef17882a4cbacf55e20c953e32d686aa63