[CrOS Cellular] Fix and run SimLockPolicyLockSetting* on more boards.
Previously, SimLockPolicyLockSetting* tests ran only on one board [1]
It is safe to run these on more boards, so this CL updates the
attributes.

This CL also modifies SimLockPolicyLockSettingOff so that it does
not fail if the primary network is not in the Connected state, but
in another state like "Sign in to connect".

[1] https://stainless.corp.google.com/search?view=matrix&row=board_model&col=build&first_date=2022-10-08&last_date=2022-10-26&test=SimLockPolicyLockSetting&build=%5ER108%7C%5ER109&exclude_cts=true&exclude_not_run=true&exclude_non_release=true&exclude_au=true&exclude_acts=true&exclude_retried=true&exclude_non_production=false&exclude_non_critical=false

BUG=b:250043471
TEST=network.SimLockPolicyLockSettingOff leasing out devices and running remotely

Change-Id: I4f3c8e0e4aaecdf985c59bb63468d63af6d8c983