ui: Wait 5sec longer for MeetCUJ WebRTC histograms
Sometimes one or two video streams are missing from WebRTC.Video.*
histograms in a MeetCUJ test. I don't know a reliable way to reproduce
this. As a speculative fix, allow a little extra time for the histograms
to be recorded after the video stream info has disappeared from
chrome://webrtc-internals.

BUG=b:255342950
TEST=tast run <DUT> ui.MeetCUJ.16p

Change-Id: I7a5ae695516f20fa24e63bfa9b8d6f829bc4c032