tast: anticipate upcoming multiword change
As mentioned in the linked bug, some new criteria is being added to the
triggering logic of multi word suggestions. Now the preceding text to a
multiword suggestion must be at least 2 or more characters.

In anticipation of this change, this CL updates the relevant multiword
tast tests to only include text examples that have at least 2 or more
characters to ensure they still generate suggestions, and thus pass.

BUG=b:255679709
TEST=tast -verbose run -build=true 127.0.0.1:9222
inputs.PhysicalKeyboardMultiwordsuggestion

Change-Id: Ib2f841ffa7485d7acc1933fecec616e36e7d29c8