rollback: Remove milestone check <100 from tests
Some verification steps in rollback are expected to fail if the image
version installed is <100. We skip this verification so the tests do
not fail when they roll back to an image M<100.

We do not allow rolling back to such old images or test them anymore.

Remove the milestone check from tests as it is no longer needed.

BUG=b:237500398
TEST=tast -verbose run DUT policy.EnterpriseRollbackInPlace

Change-Id: I124be5ffc02348f13de1879ce70dee9318c9ada5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3986085
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>