Reduce SAMLLogin flakiness by waiting for SAML paging to be loaded
performGAIALogin() is now waiting for the SAML page to be loaded
instead of directly passing the context to the tast test to reduce
test flakiness for the DeviceTrustLoginScreen tast test as well as
for other tast tests using SAMLLogin as LoginMode.
Add screenshot functionality on the login screen.

BUG=b:255713782
TEST=tast run <ip> policy.DeviceTrustLoginScreen.*

Change-Id: I01b83aa12267a779c96407d4bb2e0c21bfe8c4df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3987603
Reviewed-by: Mohamed Omar <mohamedaomar@google.com>
Tested-by: Leon Masopust <lmasopust@google.com>
Commit-Queue: Leon Masopust <lmasopust@google.com>
Reviewed-by: Martin Rodriguez <rodmartin@google.com>