ui: Lacros variants to ui.HotseatAnimation
This test needs lacros variants since it interacts with browser.
The metrics are recorded in ash-chrome.

For the WindowCreation subtests, it changes to using apps.Launch
to be able to open a new browser window that works consistently
for both ash and lacros-chrome.

Also stabilize ash.EnterShelfOverflow to wait for the lacros or
ash extension apps to be installed. For Lacros it would take a few more
seconds for ash.ChromeApps to return all the installed extensions apps.

BUG=b:216242056
TEST=tast run ui.HotseatAnimation*

Change-Id: I5c84d2399c23720445dc7796005f6b8ed8ff4a68