tast-tests: Add new server features to tast tests
Changing the ReportingUtil to leverage the new features that we implemented on the server. We can now query specifically for one ClientId and filter by time directly from the query to the server instead of doing it locally.

Updated the LookupEvents to query the server this way and deleted the logic from the PruneEvents func and others that were handling it locally.

TEST=tast -verbose run DUT policy.HeartbeatReporting*

Bug: b:238405296
Change-Id: I6db13dad318b982b9f1be2159043b192293a681d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3989283
Tested-by: Alberto Juarez <albertojuarez@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Alberto Juarez <albertojuarez@google.com>