tast-tests: Add some verification tests in a VerifyFactors test.
This adds a new test that's focused on testing the verify paths for
different factors. Initially this one just starts with a password auth
factor but we will add support for more of them in subsequent changes.

The test is parameterized on both VK vs USS and persistent versus
ephemeral users.

For reusability I added a WithAuthSession utility function to the
cryptohome client library. This lets you execute a block of operations
within an auth session, handling a bunch of the setup and teardown
boilerplate for you. This was based on a suggestion from an earlier
change, and upon use it does seem to make the test code a little less
noisy.

BUG=b:255985000
TEST=tast run <dut> cryptohome.VerifyFactors.*

Change-Id: Iec941902fc3b5a983600530c2d4d4e1caa36d826