tast-tests: Make AuthSessionUnlock into a more general verify test.
This reworks the test to add a couple of new features: it checks both
persistent and ephemeral users (the existing version just tests
persistent users) and it also includes a password change.

The password change makes it overlap somewhat with UpdatePassword, but
I'm not sure testing an unlock is really complete without the update,
which at a bare minimum allows us to confirm that verifiers don't
persist after the underlying credentials that produced them were
removed.

The test is also significantly refactored in order to utilize our test
fixture for setting up USS vs VKs. It also makes use of a new utility I
added for reusability, a WithAuthSession utility function in the
cryptohome client library. This lets you execute a block of operations
within an auth session, handling a bunch of the setup and teardown
boilerplate for you. This was based on a suggestion from an earlier
change, and upon use it does seem to make the test code a little less
noisy.

Note that the end result of the refactoring is that in practice it is
less an update of AuthSessionUnlock and more of a rewrite of it.

BUG=b:255985000
TEST=tast run <dut> cryptohome.AuthSessionUnlock.*

Change-Id: Iec941902fc3b5a983600530c2d4d4e1caa36d826
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3989744
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Judy Wu <jinrongwu@google.com>